+++ src/SMAPI/Framework/Events/ModSpecialisedEvents.cs
@@ -9,11 +9,11 @@ namespace StardewModdingAPI.Framework.Events
         /*********
         ** Accessors
         *********/
-        /// <summary>Raised immediately after the player loads a save slot, but before the world is fully initialised. The save and game data are available at this point, but some in-game content (like location maps) haven't been initialised yet.</summary>
-        public event EventHandler<SavePreloadedEventArgs> SavePreloaded
+        /// <summary>Raised when the low-level stage in the game's loading process has changed. This is an advanced event for mods which need to run code at specific points in the loading process. The available stages or when they happen might change without warning in future versions (e.g. due to changes in the game's load process), so mods using this event are more likely to break or have bugs. Most mods should use <see cref="IGameLoopEvents"/> instead.</summary>
+        public event EventHandler<LoadStageChangedEventArgs> LoadStageChanged
         {
-            add => this.EventManager.SavePreloaded.Add(value);
-            remove => this.EventManager.SavePreloaded.Remove(value);
+            add => this.EventManager.LoadStageChanged.Add(value);
+            remove => this.EventManager.LoadStageChanged.Remove(value);
         }
 
         /// <summary>Raised before the game state is updated (≈60 times per second), regardless of normal SMAPI validation. This event is not thread-safe and may be invoked while game logic is running asynchronously. Changes to game state in this method may crash the game or corrupt an in-progress save. Do not use this event unless you're fully aware of the context in which your code will be run. Mods using this event will trigger a stability warning in the SMAPI console.</summary>